refactor: 폴더,파일  네이밍 변경 및 리팩토링
