+++ src/pages/signup/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
-import Carousel from '@/components/loginLogout/Carousel';
-import SignupForm from '@/components/loginLogout/SignupForm';
+import Carousel from '@/components/loginSignup/Carousel';
+import SignupForm from '@/components/loginSignup/SignupForm';
 import { SEO } from '@/components/shared/SEO';
 import useInternalRouter from '@/hooks/useInternalRouter';
 import { userStore } from '@/store/userStore';
@@ -17,21 +17,21 @@ export default function SignupPage() {
 
   useEffect(() => {
     if (isLoggedIn) {
-      void router.push('/main'); // 예시로 대시보드 페이지로 리디렉션
+      void router.push('/main');
     }
   }, [isLoggedIn, router]);
 
   return (
     <>
       <SEO title="manchui-signup" />
-      <div className="flex h-screen bg-white pc:w-full pc:bg-white">
-        <div className="flex w-full flex-col items-center justify-center pc:w-1/2 pc:space-y-6">
+      <main className="flex h-screen bg-white pc:w-full pc:bg-white">
+        <section className="flex w-full flex-col items-center justify-center pc:w-1/2 pc:space-y-6">
           <SignupForm />
-        </div>
-        <div className="relative hidden h-full w-1/2 flex-col items-center justify-center bg-blue-800 pc:flex">
+        </section>
+        <aside className="relative hidden h-full w-1/2 flex-col items-center justify-center bg-blue-800 pc:flex">
           <Carousel />
-        </div>
-      </div>
+        </aside>
+      </main>
     </>
   );
 }
